backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
python-dotenv==1.1.0
requests==2.32.3
pydantic==2.11.3
google-generativeai==0.8.5
Pillow==11.2.1
pytesseract==0.3.13
python-multipart==0.0.20 